+++ NEWS.md
@@ -159,6 +159,8 @@
 
 10. The `datatable.old.unique.by.key` option has been warning for 1 year that it is deprecated: `... Please stop using it and pass by=key(DT) instead for clarity ...`. This warning is now upgraded to error as per the schedule in note 10 of v1.11.0 (May 2018), and note 1 of v1.9.8 (Nov 2016). In June 2020 the option will be removed.
 
+11. There is new message informing (once per session) about global options that should be considered unsafe, namely `datatable.nomatch` and `datatable.naturaljoin`. Please provide `nomatch` argument and `on=.NATURAL` explicitly instead of using options, [#3585](https://github.com/Rdatatable/data.table/issues/3585).
+
 
 ### Changes in [v1.12.2](https://github.com/Rdatatable/data.table/milestone/14?closed=1)  (07 Apr 2019)